requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]